requirements.txt
astroid==2.3.2
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
emoji==0.5.4
future==0.18.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
pylint==2.4.3
PySocks==1.7.1
pyTelegramBotAPI==3.6.6
python-telegram-bot==12.2.0
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.6
wrapt==1.11.2
clarifai